create put_mem

+++ printf/print_lst.c
@@ -6,7 +6,7 @@
 /*   By: junhseo <junhseo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/05 17:24:11 by junhseo           #+#    #+#             */
-/*   Updated: 2023/04/07 18:54:33 by junhseo          ###   ########.fr       */
+/*   Updated: 2023/04/07 18:57:59 by junhseo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	print_lst(const char type_spec, va_list *lst, int *count)
 	else if (type_spec == 's')
 		put_str(va_arg(*lst, char *), count);
 	else if (type_spec == 'p')
-		put_mem(va_arg(*lst, unsigned int), count);
+		put_mem(va_arg(*lst, unsigned long long), count);
 	else if (type_spec == 'd' || type_spec == 'i')
 		put_nbr(va_arg(*lst, int), count);
 	else if (type_spec == 'u')
